delete

+++ delete.php
@@ -2,22 +2,32 @@
 require 'db_connect.php';
 
 if(isset($_POST['delete'])){
+    try{
+        $del = $_POST['delete'];
 
-    $del = $_POST['delete'];
+        if(empty($text)){
+            $message = "Veuillez mettre du text";
+            echo '<script type="text/javascript">window.alert("'.$message.'");</script>';
+        } else{
+            $deleteList = $conn->prepare("DELETE FROM todolist(ToDoListName) WHERE id=?");
+            $delete = $deleteList->execute([$del]);
+        }
 
-    echo $del;
-
-    $deleteList = $conn->prepare("DELETE FROM todolist(ToDoListName) WHERE id=?");
-    $delete = $deleteList->execute([$del]);
-
-    if($delete){
-        header('Location: http://php-todolist/toDoList.php');
-    } else {
-        header('Location: http://php-todolist/toDoList.php');
+        if($delete){
+            header('Location: http://php-todolist/toDoList.php');
+        } else {
+            header('Location: http://php-todolist/toDoList.php');
+        }
+    } catch(PDOException $e) {
+        echo "Connection failed: " . $e->getMessage();
     }
-} else {
-    $message = "ERROR";
-    echo '<script type="text/javascript">window.alert("'.$message.'");</script>';
 }
 
-?>
+
+// if(isset($_POST['delete'])){
+//     try{
+//         $del = $_POST['delete'];
+//         $deleteList = $conn->prepare("DELETE FROM INSERT INTO todolist(ToDoListName) WHERE id=?");
+//         $deleteList->execute([$del]);
+    
+// }
